+++ src/preview_canvas.cpp
@@ -365,8 +365,8 @@ int PreviewCanvas::Pick() {
 
 	sphere_list = glGenLists(1);
 	glNewList(sphere_list, GL_COMPILE);
-	gluSphere(quadric, 1.0f, (long) (1.5f * gPreferences->GetQD3DAtomQuality()),
-			  (long) (gPreferences->GetQD3DAtomQuality()));
+	gluSphere(quadric, 1.0f, (GLint) (1.5f * gPreferences->GetQD3DAtomQuality()),
+			  (GLint) (gPreferences->GetQD3DAtomQuality()));
 	glEndList();
 
 	glRenderMode(GL_SELECT);
